% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chain_file.R
\name{GRangesMappingToChainFile}
\alias{GRangesMappingToChainFile}
\title{Create exome chain file from genome annotation file}
\usage{
GRangesMappingToChainFile(
  input_gtf,
  out_chain_name,
  chrom_suffix = "exome",
  verbose = FALSE
)
}
\arguments{
\item{input_gtf}{The name of the gtf/gff file that will be converted to an exome chain file}

\item{out_chain_name}{The name of the chain file to be created}

\item{chrom_suffix}{The suffix to be appended to all chromosome names created in the chain file}

\item{verbose}{Whether updates are desired while the function is running}
}
\description{
`GRangesMappingToChainFile` writes a chain file of the exome mapped from a genome annotation file.
}
\examples{
\dontrun{
GRangesMappingToChainFile("hg19_annotations_ensembl.gtf", "hg19_annotations_ensembl_exome.chain", verbose=T)
}

}
