% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapRangesToGenomicIntervals.R
\name{GRangesMappingToChainViaFile}
\alias{GRangesMappingToChainViaFile}
\title{GRangesMappingToChainViaFile}
\usage{
GRangesMappingToChainViaFile(
  ranges_to_map_to,
  chrom_suffix = "_mapped",
  out_chain_name = "",
  chromosomes_length = c(),
  verbose = FALSE
)
}
\arguments{
\item{ranges_to_map_to}{A GRanges file with non-overlapping intervals that will be converted to a chain file. Required.}

\item{chrom_suffix}{The suffix to be appended to all the sestination chromosome names in the mapping "default is "_mapped"}

\item{out_chain_name}{The name of a chain file to be written in the local directory. Default is "", is calls a tmp file creation. In this case, the file will be unlinked before the function returns.}

\item{chromosomes_length}{is sequinfo of the mapping object is not enough for the chromosome lengths, the additional info is provided here. Default is c(). The foemat is like the seqlengths() result for a GRanges.}

\item{verbose}{Output updates while the function is running. Default FALSE}
}
\value{
a Chain object that maps all the chromosomes according to GRanges
}
\description{
GRangesMappingToChainViaFile creates a \code{Chain} object based on intervals from a \code{GRanges} object. The mapping by this \code{Chain} will collapse chromosomes of the annotation into pseudogenome that are combined from tiled intervals of the \code{GRanges} object.
}
